Minor touch ups to privacy of cloth mixins (#652)
